cmake: Remove -Werror
This is bad (causes random build failures based on compiler version, especially when used wiht -Wall/-Wextra),
wrong (it does not guarantee that the code is actually correct),
and dangerous (provides a dellusion that the code is OK).
The only quarantee it enforces is that the compiler cares less about the standard than the programmer.

Fixes build failure using gcc-8

Signed-off-by: Jan Vesely <jan.vesely@rutgers.edu>